Fix static setting and new_bill.html values

+++ homeland_manager/forms.py
@@ -2,9 +2,9 @@ from django import forms
 
 class MonthlyBills(forms.Form):
     cleaner_tax = forms.IntegerField(label='Cleaner tax')
-    stairs_electricity = forms.IntegerField(label='Stairs electricity')
-    elevator_electricity = forms.IntegerField(label='Elevator electricit')
-    other_common = forms.IntegerField(label='Other common payments')
-    other_elevator = forms.IntegerField(label='Other elevator payments')
+    stairs_electricity = forms.FloatField(label='Stairs electricity')
+    elevator_electricity = forms.FloatField(label='Elevator electricit')
+    other_common = forms.FloatField(label='Other common payments')
+    other_elevator = forms.FloatField(label='Other elevator payments')
     common_comment = forms.CharField(label='Common comment', max_length=10)
     elevator_comment = forms.CharField(label='Elevator comment', max_length=10)
